chore: move hero section components inside the HeroSection/components
