+++ frontend/src/container/CloudIntegrationPage/HeroSection/components/IntegrationModal.style.scss
