+++ frontend/src/container/CloudIntegrationPage/HeroSection/components/RegionSelector.tsx
@@ -4,7 +4,7 @@ import { Checkbox } from 'antd';
 import { useRegionSelection } from 'hooks/integrations/aws/useRegionSelection';
 import { Dispatch, SetStateAction } from 'react';
 
-import { regions } from '../ServicesSection/data';
+import { regions } from '../../ServicesSection/data';
 
 export function RegionSelector({
 	selectedRegions,
