+++ frontend/src/container/CloudIntegrationPage/HeroSection/components/SuccessView.style.scss
