+++ frontend/src/container/CloudIntegrationPage/HeroSection/components/SuccessView.tsx
