+++ frontend/src/hooks/integrations/aws/useAwsAccounts.ts
@@ -1,5 +1,5 @@
 import { getAwsAccounts } from 'api/integrations/aws';
-import { CloudAccount } from 'pages/Integrations/CloudIntegrationPage/ServicesSection/types';
+import { CloudAccount } from 'container/CloudIntegrationPage/ServicesSection/types';
 import { useQuery, UseQueryResult } from 'react-query';
 
 export const AWS_ACCOUNTS_QUERY_KEY = ['aws-accounts'];
